requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.0
discord-ui>=5.1.0

# Database Dependencies
SQLAlchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Web Framework (for calendar endpoints)
fastapi>=0.104.0
uvicorn>=0.24.0

# Natural Language Processing
spacy>=3.7.0
python-dateutil>=2.8.0
dateparser>=1.1.0

# Calendar Integration
icalendar>=5.0.0

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Utilities
pytz>=2023.3
aiohttp>=3.8.0
asyncpg>=0.29.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Production Dependencies
gunicorn>=21.2.0
